Updated team rankings, each team now has division rank and overall rank. Updated teamApiCall to divisionTeamApiCall.
